track/cache outline nodes per DesignContext, so we can cleanup with that DesignContext
