+++ WpfDesign.Designer/Project/PropertyGrid/Editors/CollectionEditor.xaml.cs
@@ -79,7 +79,7 @@ namespace ICSharpCode.WpfDesign.Designer.PropertyGrid.Editors
 			if(control!=null){
 				TypeMappings.TryGetValue(control.GetType(), out _type);
 				if (_type != null) {
-					IOutlineNode node = OutlineNode.Create(item);
+					IOutlineNode node = item.CreateOutlineNode();
 					Outline.Root = node;
 					PropertyGridView.PropertyGrid.SelectedItems = item.Services.Selection.SelectedItems;
 				}
